pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "matrix-gptbot"
version = "0.1.0-alpha1"

authors = [
  { name="Kumi Mitterer", email="gptbot@kumi.email" },
]

description = "Multifunctional Chatbot for Matrix"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"

packages = [
    "src/gptbot"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "matrix-nio[e2e]",
    "markdown2[all]",
    "tiktoken",
    "python-magic",
    "pillow",
    ]

[project.optional-dependencies]
openai = [
    "openai",
]

wolframalpha = [
    "wolframalpha",
]

trackingmore = [
    "trackingmore @ git+https://kumig.it/kumitterer/trackingmore-api-tool.git",
]

all = [
  "matrix-gptbot[openai,wolframalpha,trackingmore]",
]

dev = [
  "matrix-gptbot[all]",
  "black",
]

[project.urls]
"Homepage" = "https://kumig.it/kumitterer/matrix-gptbot"
"Bug Tracker" = "https://kumig.it/kumitterer/matrix-gptbot/issues"

[project.scripts]
gptbot = "gptbot:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gptbot"]